requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
scipy
pyyaml
joblib
